fix: 코드리뷰 반영
게임 플레이 전체를 관장하는 게임 객체를 생성해서 게임을 진행함.

pr: #1
